apps/fastapi-face-verification-be/requirements.txt
pandas
seaborn
matplotlib
numpy
scikit-learn
dill
fastapi
uvicorn
python-dotenv
-e .
